backends: Rename _util to _common
To include LoopAbort in another commit.
